rabbit/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.2
pika==1.3.2